fix: 이미지 불러오기 상태관리 추가
