+++ src/components/EditPage/ItemListBox/2D/ImageItem.jsx
@@ -1,18 +1,12 @@
 import styled from 'styled-components';
-import { useState } from 'react';
+import { useRecoilState } from 'recoil';
 import useImage from '../../../../hooks/AddItem/useImage';
 import ImageList from './atom/ImageList';
-
-const initialImgUrl = [
-  { url: '/logoImg/logo1.png' },
-  { url: '/logoImg/logo2.png' },
-  { url: '/logoImg/logo3.png' },
-  { url: '/logoImg/logo4.png' },
-];
+import { imageState } from '../../../../store/initialState';
 
 function ImageItem({ menuRef }) {
   const { addImage } = useImage();
-  const [uploadedImages, setUploadedImages] = useState([]);
+  const [images, setImages] = useRecoilState(imageState);
 
   const handleFileChange = (event) => {
     const file = event.target.files[0];
@@ -20,7 +14,7 @@ function ImageItem({ menuRef }) {
       const reader = new FileReader();
       reader.onload = (e) => {
         const newImage = { url: e.target.result };
-        setUploadedImages((prevImages) => [...prevImages, newImage]);
+        setImages((prevImages) => [...prevImages, newImage]);
         addImage(e.target.result);
       };
       reader.readAsDataURL(file);
@@ -29,7 +23,7 @@ function ImageItem({ menuRef }) {
 
   return (
     <ItemContainer ref={menuRef}>
-      {initialImgUrl.concat(uploadedImages).map((data, index) => (
+      {images.map((data, index) => (
         <ImageList
           key={index}
           onClick={() => addImage(data.url)}
